cdk/requirements.txt
aws-cdk-lib==2.123.0
constructs==10.3.0